requirements.txt
amqp==5.0.9
appdirs==1.4.3
asgiref==3.4.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
CacheControl==0.12.6
celery==5.2.6
certifi==2019.11.28
chardet==3.0.4
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.0
distro==1.4.0
django-cors-headers==3.11.0
django-crontab==0.7.1
Django==3.2.9
django-filter==21.1
django-rest-auth==0.9.5
django-rest-swagger==2.1.1
djangorestframework==3.12.2
drf-yasg==1.20.0
gunicorn==20.1.0
html5lib==1.0.1
humanize
idna==2.8
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.3
lockfile==0.12.2
MarkupSafe==2.0.1
mock==4.0.3
msgpack==0.6.2
openapi-codec==1.3.2
packaging==20.3
pep517==0.8.2
Pillow==9.0.0
progress==1.5
prompt-toolkit==3.0.24
psycopg2-binary==2.9.2
pyparsing==2.4.6
pytoml==0.1.21
pytz==2021.3
redis==3.5.3
django-allauth
requests==2.22.0
retrying==1.3.3
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.15.0
sqlparse==0.4.2
translate
uritemplate==4.1.1
urllib3==1.25.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
opencv-python==4.2.0.32
pytesseract
